refactor: isolate catalog migration rollout code
