+++ i18n/i18n.config.ts
@@ -2,4 +2,5 @@ export default defineI18nConfig(() => ({
     fallbackLocale: 'en',
     fallbackWarn: true,
     missingWarn: true,
+    warnHtmlMessage: false,
 }))
