+++ i18n/locales/ja-JP.json
@@ -191,8 +191,8 @@
     },
     "index": {
         "hero": {
-            "title": "お気に入りの\u200b改変を\u200b記録して\u200b共有しよう",
-            "description": "あのアイテムって\u200bどれだっけ？ \u200b記録しておけば、\u200bもう忘れません。"
+            "title": "お気に入りの<wbr>改変を<wbr>記録して<wbr>共有しよう",
+            "description": "あのアイテムって<wbr />どれだっけ？ <wbr />記録しておけば、<wbr />もう忘れません。"
         },
         "tabs": {
             "latest": "最新",
