+++ server/migration/catalog/compatibility.ts
@@ -1,6 +1,11 @@
-import { catalogItems, itemSources, publisherSources, publishers } from '@@/database/schema'
+import { catalogItems, items, itemSources, publisherSources, publishers } from '@@/database/schema'
+import type { ItemCategory } from '@avatio/core/catalog'
 import { and, eq, inArray } from 'drizzle-orm'
 import type { BatchItem } from 'drizzle-orm/batch'
+import type { AppDatabase } from '~~/server/utils/database'
+import { platformSchema, type Platform, type SetupItem } from '~~/shared/types/database'
+
+import { legacyItemCanonicalUrl, legacyPublisherCanonicalUrl } from './legacy'
 
 const SOURCE_FRESHNESS_MS = 24 * 60 * 60 * 1000
 
@@ -232,4 +237,155 @@ export const markCatalogCompatibilityWithdrawal = async (
         )
         .returning({ itemId: itemSources.itemId })
     return source?.itemId
+}
+
+const metadataNumber = (metadata: Record<string, unknown> | null, key: string) => {
+    const value = metadata?.[key]
+    return typeof value === 'number' ? value : undefined
+}
+
+const metadataString = (metadata: Record<string, unknown> | null, key: string) => {
+    const value = metadata?.[key]
+    return typeof value === 'string' ? value : undefined
+}
+
+const metadataContributors = (metadata: Record<string, unknown> | null) => {
+    const contributors = metadata?.contributors
+    if (!Array.isArray(contributors)) return undefined
+    return contributors.flatMap((contributor) => {
+        if (!contributor || typeof contributor !== 'object') return []
+        const name = Reflect.get(contributor, 'name')
+        const avatar = Reflect.get(contributor, 'avatar')
+        if (typeof name !== 'string') return []
+        return [{ name, ...(typeof avatar === 'string' ? { avatar } : {}) }]
+    })
+}
+
+interface LegacySetupItemProjectionInput {
+    source: {
+        providerKey: string
+        externalId: string
+        displayName: string
+        image: string | null
+        price: string | null
+        popularityCount: number | null
+        nsfw: boolean
+        metadata: Record<string, unknown> | null
+        publisherSource: {
+            externalId: string
+            name: string
+            image: string | null
+            providerVerified: boolean
+        } | null
+    }
+    displayNameOverride: string | null
+    category: ItemCategory
+    unsupported: boolean
+    note: string | null
+    shapekeys: SetupItem['shapekeys']
+}
+
+/**
+ * Temporary adapter for the pre-v2 client shape. Unknown providers are absent
+ * only because that legacy DTO has a closed platform union; the v2 Catalog and
+ * Setup domains do not branch on provider keys.
+ */
+export const projectCatalogSourceToLegacySetupItem = (
+    input: LegacySetupItemProjectionInput,
+): SetupItem | null => {
+    const platform = platformSchema.safeParse(input.source.providerKey)
+    if (!platform.success) return null
+
+    const publisher = input.source.publisherSource
+    return {
+        id: input.source.externalId,
+        platform: platform.data,
+        category: input.category,
+        name: input.source.displayName,
+        niceName: input.displayNameOverride,
+        image: input.source.image,
+        price: input.source.price,
+        likes: input.source.popularityCount,
+        nsfw: input.source.nsfw,
+        outdated: false,
+        shop: publisher
+            ? {
+                  id: publisher.externalId,
+                  platform: platform.data,
+                  name: publisher.name,
+                  image: publisher.image,
+                  verified: publisher.providerVerified,
+              }
+            : null,
+        unsupported: input.unsupported,
+        note: input.note,
+        shapekeys: input.shapekeys,
+        forks: metadataNumber(input.source.metadata, 'forks'),
+        version: metadataString(input.source.metadata, 'version'),
+        contributors: metadataContributors(input.source.metadata),
+    }
+}
+
+export interface LegacySetupItemWrite {
+    id: string
+    setupId: string
+    itemId: string
+    category?: ItemCategory | null
+    unsupported?: boolean
+    note?: string | null
+}
+
+export interface LegacySetupShapekeyWrite {
+    setupItemId: string
+    name: string
+    value: number
+}
+
+/** Temporary rollout dual-write bridge from legacy Setup inputs to SetupEntry rows. */
+export const mapV2SetupEntryWrites = async (
+    db: AppDatabase,
+    entries: LegacySetupItemWrite[],
+    shapekeys: LegacySetupShapekeyWrite[],
+) => {
+    if (!entries.length) return { entries: [], shapekeys: [] }
+    const externalIds = [...new Set(entries.map((entry) => entry.itemId))]
+    const [legacyItems, sources] = await Promise.all([
+        db
+            .select({ id: items.id, platform: items.platform })
+            .from(items)
+            .where(inArray(items.id, externalIds)),
+        db
+            .select({
+                itemId: itemSources.itemId,
+                providerKey: itemSources.providerKey,
+                externalId: itemSources.externalId,
+            })
+            .from(itemSources)
+            .where(inArray(itemSources.externalId, externalIds)),
+    ])
+    const providerByExternalId = new Map(legacyItems.map((item) => [item.id, item.platform]))
+    const catalogItemByExternalId = new Map(
+        sources.flatMap((source) =>
+            providerByExternalId.get(source.externalId) === source.providerKey
+                ? [[source.externalId, source.itemId] as const]
+                : [],
+        ),
+    )
+    if (entries.some((entry) => !catalogItemByExternalId.has(entry.itemId))) return null
+
+    return {
+        entries: entries.map((entry) => ({
+            id: entry.id,
+            setupId: entry.setupId,
+            itemId: catalogItemByExternalId.get(entry.itemId)!,
+            categoryOverride: entry.category ?? null,
+            unsupported: entry.unsupported ?? false,
+            note: entry.note ?? null,
+        })),
+        shapekeys: shapekeys.map((shapekey) => ({
+            setupEntryId: shapekey.setupItemId,
+            name: shapekey.name,
+            value: shapekey.value,
+        })),
+    }
 }
