+++ server/migration/catalog/queueCompatibility.ts
@@ -1,4 +1,8 @@
 import type { CacheContext } from '@cloudflare/workers-types'
+import { getCatalogCacheInvalidator, getCatalogRepository } from '~~/server/utils/catalogRuntime'
+import { useDB } from '~~/server/utils/database'
+import getItem from '~~/server/utils/getItem'
+import type { Item, Platform } from '~~/shared/types/database'
 
 /** Temporary rollout shape for messages already present in the physical queue. */
 export interface ItemRevalidationMessage {
@@ -7,6 +11,18 @@ export interface ItemRevalidationMessage {
     reason: 'setup-detail' | 'owned-avatars'
     requestedAt: string
     force?: boolean
+}
+
+export const isLegacyItemRevalidationMessage = (
+    value: unknown,
+): value is ItemRevalidationMessage => {
+    if (!value || typeof value !== 'object') return false
+    const candidate = value as Partial<ItemRevalidationMessage>
+    return (
+        typeof candidate.id === 'string' &&
+        (candidate.platform === 'booth' || candidate.platform === 'github') &&
+        (candidate.reason === 'setup-detail' || candidate.reason === 'owned-avatars')
+    )
 }
 
 export const handleItemRevalidationMessage = async (
